requirements.txt
absl-py==2.1.0
audiomentations==0.34.1
audioread==3.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
filelock==3.9.0
fonttools==4.49.0
fsspec==2023.4.0
grpcio==1.62.0
h5py==3.10.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.3
mlxtend==0.23.1
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
nltk==3.8.1
numba==0.59.0
numpy==1.24.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyvips==2.2.2
PyWavelets==1.5.0
regex==2023.12.25
requests==2.28.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
vidaug @ git+https://github.com/okankop/vidaug@1c1ddf2640fe4a9171267d64ae5e3bd70c24d54a
Werkzeug==3.0.1
zipp==3.17.0
